Refactor compileExpandableString to a separate file
